LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-spacing/tentative/parsing/text-autospace-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing text-spacing with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-4/#text-spacing-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('text-autospace', 'auto');
test_valid_value('text-autospace', 'no-autospace');
</script>
</body>
</html>
